# WIP faster cli (notably for auto completion)