--- a/deepomatic/cli/__init__.py
+++ b/deepomatic/cli/__init__.py
@@ -1,5 +1,5 @@
-from gevent.monkey import patch_all
-patch_all(thread=False, time=False, subprocess=False)  # NOQA
+# from gevent.monkey import patch_all
+# patch_all(thread=False, time=False, subprocess=False)  # NOQA
 
 from .version import __version__  # NOQA