--- a/deepomatic/cli/cmds/studio_helpers/task.py
+++ b/deepomatic/cli/cmds/studio_helpers/task.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-import gevent
+#import gevent
 
 
 class Task(object):