--- a/deepomatic/cli/input_data.py
+++ b/deepomatic/cli/input_data.py
@@ -1,5 +1,4 @@
 import os
-import cv2
 import json
 import logging