--- a/deepomatic/cli/json_schema.py
+++ b/deepomatic/cli/json_schema.py
@@ -1,4 +1,4 @@
-from jsonschema import validate, ValidationError
+#from jsonschema import validate, ValidationError
 
 
 class JSONSchemaType: