We have a really slow cli, which makes the auto-completion from #192 unusable.

`time deepo --help` takes ~900ms; completion takes more.

This WIP PR uses `python -X importtime -c 'import deepomatic.cli.cli_parser'` to diagnostic slow imports, like in https://github.com/Deepomatic/dmake/pull/478

First non-working pass, just to find what we can gain:

- gevent monkey patch: 100ms
- jsonschema: 100ms
- cv2: large
- requests: large
- git: 50ms
- deepomatic.api: 230ms
- tenacity 40ms
- tqdm: medium
- yaml: 40ms
- uuid: 10ms

`time deepo --help`: 
python 3.6.12
- before: 950ms
- after: 550ms

python 3.8.6
- before: 830ms
- after: 170ms

autocompletion speed follows these numbers.